AB#27999 - Project Info Zones - General Code Refactoring and Cleanup
