+++ applications/Unity.GrantManager/modules/Unity.Theme.UX2/src/Unity.Theme.UX2/Bundling/UnityThemeUX2GlobalScriptContributor.cs
@@ -9,9 +9,9 @@ public class UnityThemeUX2GlobalScriptContributor : BundleContributor
     {
         context.Files.Add("/themes/ux2/layout.js");
         context.Files.Add("/themes/ux2/table-utils.js");
-        context.Files.Add("/themes/ux2/zone-extensions.js");
 
         context.Files.AddIfNotContains("/libs/pubsub-js/src/pubsub.js");
+        context.Files.AddIfNotContains("/themes/ux2/zone-extensions.js");
 
         context.Files.AddIfNotContains("/libs/datatables.net-bs5/js/dataTables.bootstrap5.min.js");
         context.Files.AddIfNotContains("/libs/datatables.net-buttons/js/dataTables.buttons.min.js");
