+++ applications/Unity.GrantManager/modules/Unity.Theme.UX2/src/Unity.Theme.UX2/wwwroot/themes/ux2/zone-extensions.js
@@ -133,3 +133,218 @@
         console.table(tableData);
     }
 })(jQuery);
+
+class UnityChangeTrackingForm {
+    constructor($form, options = {}) {
+
+        this.options = {
+            modifiedClass: 'bg-primary-subtle',
+            saveButtonSelector: options.saveButtonSelector || '#saveButton', // TODO: Set default
+            ...options
+        };
+
+
+        // Expect options.form to be a jQuery object representing the form element
+        if (!$form || !($form instanceof $) || !$form.is('form')) {
+            throw new Error('UnityChangeTrackingForm requires a jQuery form object in options.form');
+        }
+
+        this.form = $form;
+        this.modifiedFields = new Set();
+        this.originalValues = {};
+        this.saveButton = $(this.options.saveButtonSelector);
+        console.log('UnityChangeTrackingForm initialized');
+    }
+
+    init() {
+        this.captureInitialValues();
+        this.addChangeHandler();
+        this.saveButton.prop('disabled', true);
+    }
+
+    captureInitialValues() {
+        this.form.find('input, select, textarea').each((_, element) => {
+            const $el = $(element);
+            const name = $el.attr('name');
+
+            if (name) {
+                // Handle different input types
+                if ($el.is(':checkbox')) {
+                    this.originalValues[name] = $el.prop('checked');
+                } else if ($el.is(':radio')) {
+                    if ($el.prop('checked')) {
+                        this.originalValues[name] = $el.val();
+                    }
+                } else {
+                    this.originalValues[name] = $el.val();
+                }
+            }
+        });
+    }
+
+    addChangeHandler() {
+        this.form.find('input, select, textarea').on('change', (e) => {
+            const $el = $(e.target);
+            const name = $el.attr('name');
+
+            if (name) {
+                this.checkFieldModified($el, name);
+            }
+        });
+    }
+
+    checkFieldModified($element, name) {
+        let currentValue;
+
+        if ($element.is(':checkbox')) {
+            currentValue = $element.prop('checked');
+        } else if ($element.is(':radio')) {
+            if ($element.prop('checked')) {
+                currentValue = $element.val();
+            } else {
+                return; // Skip radio buttons that aren't checked
+            }
+        } else {
+            currentValue = $element.val();
+        }
+
+        const originalValue = this.originalValues[name];
+
+        if (currentValue !== originalValue) {
+            this.markAsModified($element, name);
+        } else {
+            this.markAsUnmodified($element, name);
+        }
+
+        this.updateSaveButtonState();
+    }
+
+    markAsModified($element, name) {
+        this.modifiedFields.add(name);
+        $element.addClass(this.options.modifiedClass);
+        $element.attr('data-field-modified', 'true');
+        $element.removeAttr('data-field-modified');
+    }
+
+    markAsUnmodified($element, name) {
+        this.modifiedFields.delete(name);
+        $element.removeClass(this.options.modifiedClass);
+    }
+
+    updateSaveButtonState() {
+        // TODO: Include validation check if needed
+        this.saveButton.prop('disabled', this.modifiedFields.size === 0);
+    }
+
+    /**
+     * Reset tracking without changing values
+     */
+    resetTracking() {
+        this.modifiedFields.clear();
+        this.form.find('input, select, textarea').each((_, element) => {
+            const $el = $(element);
+            $el.removeClass(this.options.modifiedClass);
+            $el.removeAttr('data-field-modified');
+        });
+        this.saveButton.prop('disabled', true);
+        this.captureInitialValues();
+    }
+
+    /**
+     * Reset form to original values and clear tracking
+     */
+    resetForm() {
+        // Reset each field to its original value
+        this.form.find('input, select, textarea').each((_, element) => {
+            const $el = $(element);
+            const name = $el.attr('name');
+
+            if (name && this.originalValues.hasOwnProperty(name)) {
+                // Handle different input types
+                if ($el.is(':checkbox')) {
+                    $el.prop('checked', this.originalValues[name]);
+                } else if ($el.is(':radio')) {
+                    if ($el.val() === this.originalValues[name]) {
+                        $el.prop('checked', true);
+                    }
+                } else {
+                    $el.val(this.originalValues[name]);
+
+                    // Re-apply any special formatting (like currency masks)
+                    if ($el.hasClass('unity-currency-input') ||
+                        $el.hasClass('numeric-mask') ||
+                        $el.hasClass('percentage-mask')) {
+                        $el.maskMoney('mask', this.originalValues[name]);
+                    }
+                }
+
+                // Remove modification styling
+                $el.removeClass(this.options.modifiedClass);
+                $el.removeAttr('data-field-modified');
+            }
+        });
+
+        // Clear tracking
+        this.modifiedFields.clear();
+        this.saveButton.prop('disabled', true);
+
+        // Trigger change events for any dependent calculations
+        this.form.find('.numeric-mask, .unity-currency-input').trigger('change');
+    }
+
+    serializeZoneObject(camelCase = true, includeDisabledFields = false) {
+        return this.form.serializeZoneFieldsets(camelCase, includeDisabledFields);
+    }
+
+    serializeZoneArray(camelCase = false, includeDisabledFields = false, nested = false) {
+        return this.form.serializeZoneFieldsets(camelCase, includeDisabledFields, nested);
+    }
+}
+
+class UnityZoneForm extends UnityChangeTrackingForm {
+    constructor($form, options = {}) {
+        super($form, options);
+        this.options = {
+            ...this.options,
+            ...options
+        };
+
+        console.log('UnityZoneForm initialized');
+    }
+
+    init() {
+        super.init();
+        this.form.on('reset', () => this.resetTracking());
+        this.initializeNumericFields();
+        this.addSubmitHandler();
+    }
+
+    initializeNumericFields() {
+        $('.numeric-mask').maskMoney({ precision: 0 });
+        $('.percentage-mask').maskMoney();
+
+        $('.numeric-mask').each(function () {
+            $(this).maskMoney('mask', this.value);
+        });
+
+        $('.percentage-mask').each(function () {
+            $(this).maskMoney('mask', this.value);
+        });
+
+        $('.remove-leading-zeros').on('input', function () {
+            let inputValue = $(this).val();
+            let newValue = inputValue.replace(/^0+(?!$)/, '');
+            $(this).val(newValue);
+        });
+
+        $('.unity-currency-input').maskMoney();
+    }
+
+    addSubmitHandler() {
+        this.form.on('submit', (e) => {
+            e.preventDefault();
+            debugger;
+            this.resetTracking();
+        });
+    }
+}
