+++ applications/Unity.GrantManager/src/Unity.GrantManager.Application.Contracts/GrantApplications/IGrantApplicationAppService.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Unity.GrantManager.Comments;
+using Unity.Modules.Shared;
 using Volo.Abp.Application.Dtos;
 
 namespace Unity.GrantManager.GrantApplications
@@ -12,7 +13,7 @@ namespace Unity.GrantManager.GrantApplications
         Task<ListResultDto<ApplicationActionDto>> GetActions(Guid applicationId, bool includeInternal = false);
         Task<GetSummaryDto> GetSummaryAsync(Guid applicationId);
         Task<GrantApplicationDto> UpdateProjectInfoAsync(Guid id, CreateUpdateProjectInfoDto input);
-        Task<GrantApplicationDto> UpdatePartialProjectInfoAsync(Guid id, PartialUpdateDto<PartialUpdateProjectInfoDto> input);
+        Task<GrantApplicationDto> UpdatePartialProjectInfoAsync(Guid id, PartialUpdateDto<UpdateProjectInfoDto> input);
         Task<GrantApplicationDto> UpdateProjectApplicantInfoAsync(Guid id, CreateUpdateApplicantInfoDto input);
         Task<GrantApplicationDto> UpdateAssessmentResultsAsync(Guid id, CreateUpdateAssessmentResultsDto input);
         Task<List<GrantApplicationLiteDto>> GetAllApplicationsAsync();
